Made point x and y support float, added theta1 and theta2 labels

+++ point.py
@@ -5,13 +5,13 @@ class Point:
     >>> Point(x, y)
     '''
 
-    def __init__(self, x: int, y: int) -> None:
+    def __init__(self, x: float, y: float) -> None:
        '''Point class constructor.'''
 
-       self.x: int = x;
-       self.y: int = y;
+       self.x: float = x;
+       self.y: float = y;
 
-       self.tuple: tuple[int, int] = (self.x, self.y);
+       self.tuple: tuple[float, float] = (self.x, self.y);
 
     def __repr__(self) -> str:
         '''Printable representation of a Point object'''
@@ -21,9 +21,9 @@ class Point:
 def getMiddlePoint(*points: Point) -> Point:
     '''Function to get the middle point of all points given as arguments'''
 
-    x_sum: int = sum(map(lambda point: point.x, points));
-    y_sum: int = sum(map(lambda point: point.y, points));
-    length: float = float(len(points));
+    x_sum: float = sum(map(lambda point: point.x, points));
+    y_sum: float = sum(map(lambda point: point.y, points));
+    length: int = len(points);
 
     middlePoint: Point = Point(x_sum / length, y_sum / length);
